creation des packages Interface et Reader
